In the newsletter scheduler, just use the local timezone if it's one of the available options.
* Extract logic of building timezone options as a helper function, so it can be used by different components
* Use that function in order to map the local timezone to one of the options of the timezone drop-down field

Fixes CV2-3214.
