+++ src/app/components/team/Newsletter/NewsletterComponent.js
@@ -13,6 +13,7 @@ import SwitchComponent from '../../cds/inputs/SwitchComponent';
 import styles from './NewsletterComponent.module.css';
 import LanguagePickerSelect from '../../cds/forms/LanguagePickerSelect';
 import SettingsHeader from '../SettingsHeader';
+import { getTimeZoneOptions } from '../../../helpers';
 import { can } from '../../Can';
 import { withSetFlashMessage } from '../../FlashMessage';
 
@@ -72,7 +73,15 @@ const NewsletterComponent = ({
   const [contentType, setContentType] = React.useState(content_type || 'static');
   const [sendEvery, setSendEvery] = React.useState(send_every || ['wednesday']);
   const [sendOn, setSendOn] = React.useState(send_on || null);
-  const [timezone, setTimezone] = React.useState(send_timezone || null);
+
+  // Just use the local timezone if it's one of the options
+  const localTimezone = Intl.DateTimeFormat().resolvedOptions().timeZone;
+  let defaultTimezone = null;
+  if (getTimeZoneOptions().find(item => item.code === localTimezone || item.value === localTimezone)) {
+    defaultTimezone = localTimezone;
+  }
+  const [timezone, setTimezone] = React.useState(send_timezone || defaultTimezone);
+
   const [time, setTime] = React.useState(send_time || '09:00');
   const [scheduled, setScheduled] = React.useState(enabled || false);
   const [datetimeIsPast, setDatetimeIsPast] = React.useState(false);
@@ -537,7 +546,7 @@ const NewsletterComponent = ({
               type={contentType}
               sendEvery={sendEvery}
               sendOn={sendOn}
-              timezone={timezone || Intl.DateTimeFormat().resolvedOptions().timeZone}
+              timezone={timezone}
               time={time}
               parentErrors={errors}
               scheduled={scheduled}
